+++ GallerySelector/src/main/java/com/jms/galleryselector/data/LocalGalleryDataSource.kt
@@ -10,6 +10,7 @@ import androidx.core.os.bundleOf
 import androidx.paging.PagingData
 import androidx.paging.PagingSource
 import com.jms.galleryselector.Constants
+import com.jms.galleryselector.manager.MediaContentManager
 import com.jms.galleryselector.model.ImageEntity
 import kotlinx.coroutines.flow.Flow
 
@@ -18,7 +19,7 @@ import kotlinx.coroutines.flow.Flow
  * Local Gallery data source
  */
 internal class LocalGalleryDataSource(
-    private val context: Context,
+    private val contentManager: MediaContentManager,
     private val galleryStream: GalleryPagingStream
 ) {
     fun getLocalGalleryImages(
@@ -30,7 +31,7 @@ internal class LocalGalleryDataSource(
 
             val _page = it.key ?: page
 
-            getCursor(
+            contentManager.getCursor(
                 uri = uri,
                 offset = (_page - 1) * pageSize,
                 limit = pageSize
@@ -51,7 +52,7 @@ internal class LocalGalleryDataSource(
     }
 
     fun getImageEntity(): ImageEntity {
-        getCursor(
+        contentManager.getCursor(
             uri = MediaStore.Images.Media.EXTERNAL_CONTENT_URI,
             offset = 0,
             limit = 1,
@@ -60,53 +61,6 @@ internal class LocalGalleryDataSource(
                 return makeImageEntity(cursor = cursor)
             } else throw IllegalStateException("Cursor is empty!!")
         } ?: throw IllegalStateException("Cursor is null!!")
-    }
-
-    private fun getCursor(
-        uri: Uri,
-        offset: Int,
-        limit: Int,
-    ): Cursor? {
-        //filter deleted media contents
-        val selection = MediaStore.MediaColumns.IS_PENDING + " = ?"
-        val selectionArgs = arrayOf("0")
-        val projection = arrayOf(
-            MediaStore.Images.ImageColumns._ID,
-            MediaStore.Images.ImageColumns.TITLE,
-            MediaStore.Images.ImageColumns.DATE_MODIFIED,
-            MediaStore.Images.ImageColumns.DATA,
-            MediaStore.Images.ImageColumns.MIME_TYPE,
-            MediaStore.Images.ImageColumns.BUCKET_DISPLAY_NAME,
-            MediaStore.Images.ImageColumns.BUCKET_ID
-        )
-
-        return if (Build.VERSION.SDK_INT > Build.VERSION_CODES.Q) {
-            val selectionBundle = bundleOf(
-                ContentResolver.QUERY_ARG_OFFSET to offset,
-                ContentResolver.QUERY_ARG_LIMIT to limit,
-                ContentResolver.QUERY_ARG_SORT_COLUMNS to arrayOf(
-                    MediaStore.Files.FileColumns.DATE_MODIFIED
-                ),
-                ContentResolver.QUERY_ARG_SORT_DIRECTION to ContentResolver.QUERY_SORT_DIRECTION_DESCENDING,
-                ContentResolver.QUERY_ARG_SQL_SELECTION to selection,
-                ContentResolver.QUERY_ARG_SQL_SELECTION_ARGS to selectionArgs
-            )
-
-            context.contentResolver.query(
-                uri,
-                projection,
-                selectionBundle,
-                null
-            )
-        } else {
-            context.contentResolver.query(
-                uri,
-                projection,
-                selection,
-                selectionArgs,
-                "${MediaStore.Files.FileColumns.DATE_MODIFIED} DESC LIMIT $limit OFFSET $offset"
-            )
-        }
     }
 
     private fun makeImageEntity(cursor: Cursor): ImageEntity {
