+++ GallerySelector/src/main/java/com/jms/galleryselector/ui/GalleryScreenViewModel.kt
@@ -7,7 +7,7 @@ import androidx.paging.PagingData
 import androidx.paging.cachedIn
 import androidx.paging.map
 import com.jms.galleryselector.data.LocalGalleryDataSource
-import com.jms.galleryselector.manager.MediaContentManager
+import com.jms.galleryselector.manager.FileManager
 import com.jms.galleryselector.model.Gallery
 import com.jms.galleryselector.model.toImage
 import kotlinx.coroutines.flow.Flow
@@ -20,7 +20,7 @@ import kotlinx.coroutines.flow.update
 import java.io.File
 
 internal class GalleryScreenViewModel constructor(
-    private val contentManager: MediaContentManager,
+    private val fileManager: FileManager,
     val localGalleryDataSource: LocalGalleryDataSource
 ) : ViewModel() {
     //selected gallery ids
@@ -65,13 +65,13 @@ internal class GalleryScreenViewModel constructor(
     }
 
     fun createImageFile(): File {
-        _imageFile = contentManager.createImageFile()
+        _imageFile = fileManager.createImageFile()
         return _imageFile ?: throw IllegalStateException("File is null!!")
     }
 
     fun saveImageFile(context: Context, max : Int, autoSelectAfterCapture: Boolean) {
         if (_imageFile != null) {
-            contentManager.saveImageFile(context = context, file = _imageFile!!)
+            fileManager.saveImageFile(context = context, file = _imageFile!!)
 
             if (autoSelectAfterCapture)
                 select(image = localGalleryDataSource.getImageEntity().toImage(), max = max)
